cgroup: re-read memory limit
Recent kernel can change the value of memory limit. For example, when setting 5000000 as
the limit, the kernel might round it up to 4997120 to align page boundary. This can also happen
if the total memory is less than the set value.

+++ src/cgroup.h
@@ -235,10 +235,10 @@ namespace lrun {
             /**
              * set memory usage limit
              * @param   bytes       limit, no limit if bytes <= 0
-             * @return  0           success
+             * @return >=0          success, real memory limit
              *         <0           failed
              */
-            int set_memory_limit(long long bytes);
+            long long set_memory_limit(long long bytes);
 
             /**
              * restart cpuacct and memory max_usage_in_bytes
